Move IRIS analysis to plugin
