+++ docs/sources/gettingstarted/quickstart.py
@@ -311,10 +311,16 @@ _ = ds.plot()
 # specific regularization parameter.
 
 # %%
-iris = scp.IRIS(reg_par=[-10, 1, 12])
-K = scp.IrisKernel(ds, "langmuir", q=[-7, -1, 50])
-iris.fit(ds, K)
-_ = iris.f[-7].plot_contour(colorbar=True)
+try:
+    from spectrochempy_iris import IRIS
+    from spectrochempy_iris import IrisKernel
+
+    iris = IRIS(reg_par=[-10, 1, 12])
+    K = IrisKernel(ds, "langmuir", q=[-7, -1, 50])
+    iris.fit(ds, K)
+    _ = iris.f[-7].plot_contour(colorbar=True)
+except ImportError:
+    pass
 
 # %%
 # %% [markdown]
