+++ environments/environment_iris.yml
@@ -1,0 +1,49 @@
+# ======================================================================================
+#
+#       This file is automatically generated to be up-to-date in the master
+#       repository.
+#
+#       !!! DO NOT MODIFY !!!
+#
+#       if you need to modify dependencies you need to follow these steps:
+#
+#       - Any change in dependencies must be first reported in "pyproject.toml"
+#         at the root of this repository.
+#
+#       - Then execute :
+#          *  pre-commit run --all-files
+#
+# ======================================================================================
+
+channel_priority: strict
+channels:
+    - conda-forge
+    - spectrocat
+    - defaults
+
+dependencies:
+
+    # CORE dependencies
+    # -----------------
+    - colorama
+    - dill
+    - ipython
+    - jinja2
+    - lazy_loader
+    - matplotlib-base
+    - numpy
+    - numpydoc
+    - osqp
+    - pluggy
+    - pint
+    - pyyaml
+    - requests
+    - scikit-learn
+    - scipy
+    - tzlocal
+    - xlrd
+
+    # IRIS dependencies
+    # -----------------
+
+    - spectrochempy-iris>=0.1,<0.2
