+++ requirements/requirements_iris.txt
@@ -1,0 +1,24 @@
+# ======================================================================================
+#
+#       This file is automatically generated to be up-to-date in the master
+#       repository.
+#
+#       !!! DO NOT MODIFY !!!
+#
+#       if you need to modify dependencies you need to follow these steps:
+#
+#       - Any change in dependencies must be first reported in "pyproject.toml"
+#         at the root of this repository.
+#
+#       - Then execute :
+#          *  pre-commit run --all-files
+#
+# ======================================================================================
+
+# CORE dependencies
+# -----------------
+-r requirements.txt
+
+# IRIS dependencies
+# -----------------
+spectrochempy-iris>=0.1,<0.2
