+++ src/spectrochempy/plotting/composite/__init__.pyi
@@ -3,14 +3,12 @@
 # ruff: noqa
 
 __all__ = [
-    "iris",
     "plotbaseline",
     "plotmerit",
     "plotscore",
     "plotscree",
 ]
 
-from . import iris
 from . import plotbaseline
 from . import plotmerit
 from . import plotscore
